+++ src/CHANGES.md
@@ -1,5 +1,6 @@
 **Added:**
-* none
+* save global resources on application exit
+* save project on application exit
 
 **Changed:**
 * none